Move weather related functions into submodule
