+++ crates/valence/src/instance.rs
@@ -11,11 +11,8 @@ use num::integer::div_ceil;
 use rustc_hash::FxHashMap;
 use valence_protocol::array::LengthPrefixedArray;
 use valence_protocol::block_pos::BlockPos;
-use valence_protocol::packet::s2c::play::game_state_change::GameEventKind;
 use valence_protocol::packet::s2c::play::particle::Particle;
-use valence_protocol::packet::s2c::play::{
-    GameStateChangeS2c, OverlayMessageS2c, ParticleS2c, PlaySoundS2c,
-};
+use valence_protocol::packet::s2c::play::{OverlayMessageS2c, ParticleS2c, PlaySoundS2c};
 use valence_protocol::sound::Sound;
 use valence_protocol::text::Text;
 use valence_protocol::types::SoundCategory;
@@ -434,38 +431,6 @@ impl Instance {
     pub fn set_action_bar(&mut self, text: impl Into<Text>) {
         self.write_packet(&OverlayMessageS2c {
             action_bar_text: text.into().into(),
-        });
-    }
-
-    /// Sends the begin rain event to all players in the instance.
-    pub fn begin_raining(&mut self) {
-        self.write_packet(&GameStateChangeS2c {
-            kind: GameEventKind::BeginRaining,
-            value: f32::default(),
-        });
-    }
-
-    /// Sends the end rain event to all players in the instance.
-    pub fn end_raining(&mut self) {
-        self.write_packet(&GameStateChangeS2c {
-            kind: GameEventKind::EndRaining,
-            value: f32::default(),
-        });
-    }
-
-    /// Sends the set rain level event to all players in the instance.
-    pub fn set_rain_level(&mut self, level: f32) {
-        self.write_packet(&GameStateChangeS2c {
-            kind: GameEventKind::RainLevelChange,
-            value: level.clamp(WEATHER_LEVEL_MIN, WEATHER_LEVEL_MAX),
-        });
-    }
-
-    /// Sends the set thunder level event to all players in the instance.
-    pub fn set_thunder_level(&mut self, level: f32) {
-        self.write_packet(&GameStateChangeS2c {
-            kind: GameEventKind::ThunderLevelChange,
-            value: level.clamp(WEATHER_LEVEL_MIN, WEATHER_LEVEL_MAX),
         });
     }
 }
