+++ crates/valence/src/instance/weather.rs
@@ -1,4 +1,8 @@
 use bevy_ecs::prelude::*;
+use valence_protocol::packet::s2c::play::game_state_change::GameEventKind;
+use valence_protocol::packet::s2c::play::GameStateChangeS2c;
+
+use super::Instance;
 
 pub const WEATHER_LEVEL_MIN: f32 = 0_f32;
 pub const WEATHER_LEVEL_MAX: f32 = 1_f32;
@@ -16,4 +20,38 @@ pub struct Weather {
     ///
     /// The [`None`] value means no thunder event.
     pub thunder: Option<f32>,
+}
+
+impl Instance {
+    /// Sends the begin rain event to all players in the instance.
+    pub fn begin_raining(&mut self) {
+        self.write_packet(&GameStateChangeS2c {
+            kind: GameEventKind::BeginRaining,
+            value: f32::default(),
+        });
+    }
+
+    /// Sends the end rain event to all players in the instance.
+    pub fn end_raining(&mut self) {
+        self.write_packet(&GameStateChangeS2c {
+            kind: GameEventKind::EndRaining,
+            value: f32::default(),
+        });
+    }
+
+    /// Sends the set rain level event to all players in the instance.
+    pub fn set_rain_level(&mut self, level: f32) {
+        self.write_packet(&GameStateChangeS2c {
+            kind: GameEventKind::RainLevelChange,
+            value: level.clamp(WEATHER_LEVEL_MIN, WEATHER_LEVEL_MAX),
+        });
+    }
+
+    /// Sends the set thunder level event to all players in the instance.
+    pub fn set_thunder_level(&mut self, level: f32) {
+        self.write_packet(&GameStateChangeS2c {
+            kind: GameEventKind::ThunderLevelChange,
+            value: level.clamp(WEATHER_LEVEL_MIN, WEATHER_LEVEL_MAX),
+        });
+    }
 }
